+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-# v0.2.1 (2014-10-01)
+# v0.2.2 (2018-10-09)
+- Fix implementation of Diff == (and added hashcode)
+- Upgrade from unittest to test package
 
+# v0.2.1 (2014-10-01)
 - Update docs to use proper markdown for dartdocs.org